+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - 🌟 `routes`: 路由转换使用`ast`树直接转换
 - uiw-admin/config
   - 🐞 修改 `publicPath` 默认值为 `"/"`
+- uiw-admin/components
+  - 🌟 `ProForm` 增加colProps api
 
 ## 5.2.21